Fix point_inside_polygon
The way I was checking whether a line ended on the horizontal wasn't
actually solving the problem. Now, check that it's the minimum for all
lines.
If they both go above the line, it'll be two Trues, and if they both go
below, it'll be two Falses, which will cancel out.